Archive/requirements.txt
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
qwen-vl-utils>=0.0.1
pypdf>=3.0.0
langchain-community>=0.0.20
atlassian-python-api>=3.41.0
markdownify>=0.11.0
accelerate>=0.20.0
